backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
django-stubs==5.1.2
django-stubs-ext==5.1.2
djangorestframework==3.15.2
djangorestframework-stubs==3.15.2
gunicorn==23.0.0
hvac==2.3.0
idna==3.10
MarkupSafe==3.0.2
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyotp==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
